requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-array-field-select==0.2.0
django-arrayfields==0.1.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-resized==1.0.2
django-sendsms==0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
optional-django==0.3.0
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyotp==2.8.0
pytz==2023.3
react==4.3.0
reportlab==4.0.4
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
stripe==5.4.0
twilio==8.2.2
typing-extensions==4.6.1
urllib3==2.0.2
yarl==1.9.2
